Refactor to blog.getPages
