+++ lib/server/generate.js
@@ -9,6 +9,7 @@ async function execute() {
   require('../write-translations.js');
 
   const metadataUtils = require('./metadataUtils');
+  const blog = require('./blog');
   const docs = require('./docs');
 
   const CWD = process.cwd();
@@ -177,29 +178,14 @@ async function execute() {
       const targetFile = join(buildDir, 'blog', filePath);
       writeFileAndCreateFolder(targetFile, str);
     });
-  // create html files for all blog pages (collections of article previews)
-  const BlogPageLayout = require('../core/BlogPageLayout.js');
-  const perPage = 10;
-  for (let page = 0; page < Math.ceil(MetadataBlog.length / perPage); page++) {
-    const language = 'en';
-    const metadata = {page, perPage};
-    const blogPageComp = (
-      <BlogPageLayout
-        metadata={metadata}
-        language={language}
-        config={siteConfig}
-      />
-    );
-    const str = renderToStaticMarkupWithDoctype(blogPageComp);
 
-    const targetFile = join(
-      buildDir,
-      'blog',
-      page > 0 ? `page${page + 1}` : '',
-      'index.html'
-    );
-    writeFileAndCreateFolder(targetFile, str);
-  }
+  // create html files for all blog pages (collections of article previews)
+  const blogPages = blog.getPages(MetadataBlog.length, siteConfig);
+  Object.keys(blogPages).forEach(pagePath => {
+    const targetFile = join(buildDir, 'blog', pagePath);
+    writeFileAndCreateFolder(targetFile, blogPages[pagePath]);
+  });
+
   // create rss files for all blog pages, if there are any blog files
   if (MetadataBlog.length > 0) {
     let targetFile = join(buildDir, 'blog', 'feed.xml');
