+++ lib/server/server.js
@@ -10,6 +10,7 @@
 function execute(port, options) {
   const extractTranslations = require('../write-translations');
   const metadataUtils = require('./metadataUtils');
+  const blog = require('./blog');
   const docs = require('./docs');
   const env = require('./env.js');
   const express = require('express');
@@ -180,30 +181,7 @@ function execute(port, options) {
     reloadMetadataBlog();
     // Generate all of the blog pages.
     removeModuleAndChildrenFromCache(join('..', 'core', 'BlogPageLayout.js'));
-    const BlogPageLayout = require(join('..', 'core', 'BlogPageLayout.js'));
-    const blogPages = {};
-    // Make blog pages with 10 posts per page.
-    const perPage = 10;
-    for (
-      let page = 0;
-      page < Math.ceil(MetadataBlog.length / perPage);
-      page++
-    ) {
-      const language = 'en';
-      const metadata = {page, perPage};
-      const blogPageComp = (
-        <BlogPageLayout
-          metadata={metadata}
-          language={language}
-          config={siteConfig}
-        />
-      );
-      const str = renderToStaticMarkupWithDoctype(blogPageComp);
-
-      const pagePath = `${page > 0 ? `page${page + 1}` : ''}/index.html`;
-      blogPages[pagePath] = str;
-    }
-
+    const blogPages = blog.getPages(MetadataBlog.length, siteConfig);
     const parts = req.path.toString().split('blog/');
     // send corresponding blog page if appropriate
     if (parts[1] === 'index.html') {
